Finish static content layout. Practice JavaScript.
The static layout for the game is relatively simple, and should not change much from this point onward.
Now, I must implement the game in JavaScript, and I know very little about JavaScript.
This will be a journey, but hopefully it ends up as the most bare-bones implementation there can be.